(#66) [cosmetic] only unlink if link was created previously => suppressing ugly error message

+++ src/common/scripts/execute_testsuite.sh
@@ -68,10 +68,9 @@ if [ -z "$GIT_URL" ]; then
 fi
 
 # unlink global node_modules in ${SAKULI_EXECUTION_DIR}
-GLOBAL_NODE_MODULES_PATH=$(npm root -g | head -n 1)
 [[ $DEBUG == true ]] && echo "Unlink global node_modules from ${SAKULI_EXECUTION_DIR}/${SAKULI_SUITE_NAME}."
-unlink ${SAKULI_EXECUTION_DIR}/${SAKULI_SUITE_NAME}/node_modules
+[ -L "${SAKULI_EXECUTION_DIR}/${SAKULI_SUITE_NAME}/node_modules" ] && unlink ${SAKULI_EXECUTION_DIR}/${SAKULI_SUITE_NAME}/node_modules
 [[ $DEBUG == true ]] && echo "Unlink global node_modules from ${SAKULI_EXECUTION_DIR}."
-unlink ${SAKULI_EXECUTION_DIR}/node_modules
+[ -L "${SAKULI_EXECUTION_DIR}/node_modules" ] && unlink ${SAKULI_EXECUTION_DIR}/node_modules
 
 exit ${SAKULI_RETURN_CODE}
